requirements.txt
altgraph==0.17
certifi==2020.12.5
chardet==4.0.0
docutils==0.16
future==0.18.2
idna==2.10
iso8601==0.1.14
Kivy==2.0.0
Kivy-Garden==0.1.4
kivy-garden.graph @ file:///home/bello/Documents/programs/kivy/kivy_garden.graph-0.4.0.tar.gz
kivymd @ https://github.com/kivymd/KivyMD/archive/master.zip
Pillow==8.1.0
pyenchant==3.2.0
Pygments==2.7.3
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.11
pyserial==3.5
PyYAML==5.4.1
requests==2.25.1
urllib3==1.26.2
